Add support for post-rewrite hooks
Add the PostRewrite hook context and base hook so that post-rewrite
hooks can be written.

We've purposely not implemented the reading of the lines passed to
STDIN, as it's not clear what kind of API would be useful here. Since
the only use of this right now is as a trigger for re-indexing tags,
we'll pass on implementing reading STDIN until a use case arises.

Change-Id: I6047c820e5f792d1bab5ef56971f9403e3cf022e
Reviewed-on: http://gerrit.causes.com/46988
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>