Change JsonSyntax hook to use hook messages
This simplifies this hook by changing it to not have to explicitly check
if there were any messages before returning pass/fail. It now simply
returns the messages and the framework will do the rest.

To support this, the MessageProcessor was updated to support hook
messages without line numbers. The theory is that messages without line
numbers should apply to the entire file.

Change-Id: I20f293cd0b741ad9879da9a738cc083ede1dc088
Reviewed-on: http://gerrit.causes.com/47212
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>