Add Overcommit::Utils::Version helper class
This will be used in an upcoming commit to support easy switching of
behavior based on git version. It piggybacks on `Gem::Version` to give
us easy version comparisons without any effort on our part, and defines
a shortcut helper so we don't need to explicitly convert version strings
to objects in order to compare.