Change hooks to be disabled by default
While all hooks that are opt-in were explicitly disabled in 1d5dd4a8,
the implementation of `Configuration#hook_enabled?` would still default
to return `true` if a hook was not explicitly enabled.

Since we want all hooks going forward to be opt-in, we need to protect
ourselves in the relatively-likely event we forget to include an
explicit `enabled: false` in a new hook's configuration, so that we
don't accidentally introduce a new automatically enabled hook.

Thus we change the implementation to default to false. While this means
we could technically omit the explicit `enabled: false` in the default
configuration, these serve as useful documentation and should best be
kept explicit.