Inject Logger into ConfigurationLoader
In order to support displaying warnings about configuration, we need to
have access to the `Logger` instance.

Inject a `Logger` into the `ConfigurationLoader` constructor and update
all uses of it to provide this dependency.

A follow-up commit will use this to display warnings for hook
configurations that don't explicitly have the `enabled` option set.