Display warning if `enabled` option not specified
Now that `enabled` will default to `false` for all except a few hooks,
we want to warn users when they haven't explicitly specified `enabled`
in their configuration.

The dependency injection here isn't pretty, but it seemed the better
option when compared to performing the check in the
`ConfigurationLoader`. Best to keep validation in the
`ConfigurationValidator`.