Fix CaseConflicts handling to work with larger file sets
When running `overcommit --run` with a large set of files,
`CaseConflicts` would call `git ls-tree` with an argument list too large
for the OS to handle. This was exacerbated by the fact that we were
passing the same directory multiple times since we were not returning
the unique set of directories.