Display coverage stats when developing locally
The Coveralls gem doesn't display coverage statistics directly in the
console when run locally. Get around this by using SimpleCov directly
when not in Travis so we get the actual coverage stats rendered in the
`coverage` directory, making it quicker to write tests and see how
much coverage has changed.