Add support for splitting long lists of arguments
When dealing with repositories with a large number of files, a problem
that sometimes appears is a long list of files being passed to a
particular command. Since operating systems have different limitations
around how long a single command can be, we need a way to split up long
commands into multiple smaller commands.

`xargs` is a tool that does this wonderfully, but requires that the tool
be explicitly installed on systems that don't have it (like Windows).
To implement a cross-platform solution, we need to provide our own (much
simpler) implementation of `xargs` so the `execute` utility helper can
transparently split up long argument lists without hook authors needing
to know the details.