Add gemfile option for setting Bundler Gemfile
There is a real need for specifying which versions of a gem Overcommit
hooks should use during the hook run, especially when those versions are
not yet released or older versions that what is currently on the machine
(since Ruby will always load the latest version by default).

Add support for a `gemfile` option which allows the developer to specify
the explicit list of gems and their respective versions to use during
hook runs.