Add integration test for gemfile config option
Since this is likely going to be an important feature, add an
integration test that ensures it is working as intended.

The main reason this is difficult is because we are already running the
tests within a Bundler context, so we need to construct an artificial
environment that still exposes Overcommit as a gem without having to
have it installed ahead of time.