Change ad hoc hooks to work like normal git hooks
The first draft of ad hoc hooks in 1f5f5b3 was rather limited: you could
only execute a command, with no access to the original arguments passed
to the original hook.

It seems like the more likely use case for this feature is to aid
developers in migrating their existing hooks over to Overcommit without
having to rewrite them. With that in mind, augment the hook
implementation to execute the command and pass the original arguments
and standard input stream.

This introduces a security vulnerability where these hooks are not
checked by the hook signing functionality, and so it's possible for an
attacker to change an existing script (e.g. post-checkout) without it
being caught by the signature checking. This will be fixed in a
follow-up commit.