Gracefully handle closed input streams in Subprocess.spawn
Tests were failing on JRuby due to the STDIN pipe being closed before we
got a chance to write to it (this was non-deterministic).

A simple workaround is for us to silently fail in this case and assume
all is well, since it's clear the spawned process is ignoring the
standard input stream anyway.