Check that hook scripts are tracked by Git
This is a stronger check than merely ensuring that they are a relative
path.

In order to work in the case where `HEAD` was not yet defined for a
repo, we had to add a repo helper `GitRepo.tracked?` which uses
`ls-files` with a specific path to determine if it is tracked.