Sign configuration in addition to hook plugins
Unfortunately, the hook signing code did not work as advertised. Since
we weren't signing the configuration file itself, an attacker could
change the value of `verify_plugin_signatures` to `false` at the same
time they added some malicious code, thereby bypassing the check
altogether.

Work around this by signing the configuration file itself. In order to
force everyone's attention (and also to reflect the fact that this
signing doesn't only apply to hooks) change the option name to
`verify_signatures`. The default is `true`, so developers will have to
update their configuration to the new `verify_signatures` option.
Hopefully this will prompt organizations to read the change log and
perform an audit of their repository hooks.

This implementation is a little bit tangly, but I erred on the side of
having it more global in scope so that any disabling of verification is
intentional rather than because we forgot to enable it for a particular
code path.