Fix BundleCheck hook to work with --run flag with local changes
As reported in #293, this hook would fail if you ran Overcommit via the
`--run` flag with local changes that weren't staged, resulting in `git
diff -- Gemfile.lock` always exiting unsuccessfully.

Work around this by reading the contents of the file before and after,
and raise an error if its content changes.