Add an `included_files` method to base hook
In the minitest pre-push hook, we don't want to use `applicable_files`
but run the whole suite instead. However, we do want to allow
expliciting an `include` pattern in the configuration. Extracting an
`included_files` method so it can be reused in the hook seemed the
easiest solution.

This commit includes some minor refactorings.