Rename MsgPattern hook to MessageFormat
A bit nit-picky, but we've already established a convention of using
"Message" instead of the "Msg" short form (CommitMsg doesn't count as it
is a separate naming scheme brought over from Git). Also, I personally
find "Format" a better descriptor than "Pattern", as the later implies
some sort of repetition, where the former is simply "the way in which
something is arranged or set out." (definition)

I updated the implementation to not use instance variables (since there
was no need).