Fix jruby build (#467)
* Use older version of bundler for jruby builds

The jruby build started failing when Travis started using bundler
version 1.14.3. It all worked on 1.13.7 so I'm switching back to that
version when a jruby build is detected.

[1] https://travis-ci.org/brigade/overcommit/jobs/194979683
[2] https://travis-ci.org/brigade/overcommit/jobs/189188271

* Pre-install rainbow to work around bundler issues on jruby

Installing rainbow manually before doing bundle install helped fix the
build locally for me.

* Apply workaround for failing rainbow install

I think I've found an issue describing the issues we've seen with the
jruby build:
https://github.com/sickill/rainbow/issues/48

A workaround that people have applied is to run `gem update --system` as
a before_install hook. I'm giving that a try.

* Only apply gem update workaround for jruby

The workaround added in the previous commit seems to be working, but
only for the jruby build. We can conditionally apply it to prevent
failing builds for other ruby versions.