Auto-sign config file when installing hooks the first time
Previously, we required users to explicitly run `overcommit --sign` in
their repositories after running `overcommit --install`. This was
slightly annoying and made for a poor user onboarding experience.

For all practical purposes, the initial bootstrap case seems like one
where ease of getting started with the tool trumps the risk of malicious
code execution via hooks. It is reasonable to assume that a user who
runs `overcommit --install` is accepting responsibility for hook code
executed as of that point. It is only on subsequent changes to
configuration/hooks that they should be warned to sign the config
file/hooks appropriately.