Gracefully handle gem loading errors
The relaxing of the `childprocess` gem dependency constraint in 89dc99a8
resulted in us seeing the following error running `overcommit --run` in
the local Overcommit repository:

    .../lib/bundler/runtime.rb:317:in `check_for_activated_spec!':
    You have already activated childprocess 0.6.3, but your Gemfile
    requires childprocess 0.7.0. Prepending `bundle exec` to your
    command may solve this. (Gem::LoadError)

Since we didn't prepend `bundle exec` to the command, the version of
`overcommit` that is executed is whichever is the latest version
installed in the local set of installed gems. In the example above, I
had Overcommit 0.39.1 installed, which depends on childprocess ~> 0.6.3.
Thus `childprocess` 0.6.3 is loaded, and when we attempt to invoke
`Bundler.setup` we get the `Gem::LoadError` referenced above since the
local `Gemfile.lock` expects `childprocess` 0.7.0.

We can work around this by catching this specific error and re-execing
with `bundle exec` prepended to the command.