Update Pre-Push Context Modified Files
Previously, `Overcommit::HookContext::PrePush::modified_files` is set as `[]` which
does not give us any value in `applicable_files` as in `PreCommit` context.

This change updates the method to return list of added and updated files:

1. compared between current branch and tracking branch if tracking branch is available.

2. compared between current branch with parent branch if tracking branch is not available.

This gives us above values in `applicable_files` in `PrePush` context.

One usage example is we can create a hook to scan prohibited keywords such as `eval`, `console.log`, etc. before pushing to remote. Previously we can only do it in `PreCommit` hook but now we could also do it in `PrePush`.