Allow ad hoc hooks to run executations not tracked by Git
Attackers don't have access to modify executables in the user's `PATH`
(unless the user included a directory in the repository itself in their
`PATH`).

In order to make it easier to write hooks, allow them to reference
executables in the `PATH`.

If an attacker changes the command, the signature will change, so we
don't need to worry about `rm -rf /` and the like.