Bug/650 traceback when setting parallelize to false (#656)
* Reproduce deadlock when parallize is false

* Fix potential deadlock when waiting for hooks

If a hook is running while another hook tries to start and there are
not enough slots left then it will wait for the first hook to stop and
release its slots. This first hook will not signal the second if there
are no more hooks left. This results in a deadlock.

This situation is far more likely to occur when parallelize is false
since the hook will consume all the available slots.