Avoid stashing changes when all changes are staged
It was suggested in #135 that a potential way to avoid unnecessary file
change detection and general stashing woes would be to stash changes
only when there were unstaged changes that needed to be excluded from
the run.

This presents a nice optimization that I wish we had considered long
ago. We can simplify the logic now that we see the purpose of this
feature a little more clearly: stash only when you need to.

Issue #669 raised an interesting point about the difficulty to detect if
a stash entry was created, so we modified the tests to check if a method
call was made. This isn't great as it's peeking deep into the
implementation, so we added some additional tests that ensured the
methods in question actually existed. Not perfect, but gives us some
confidence in the implementation.

While here, we were able to:

* Extract the stashing logic into a separate `stash_changes` helper
  method.

* Simplify the `cleanup_environment` logic to just use whether changes
  were stashed in order to decide whether to restore.

Closes #669.